# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "![Smart](http://localhost:8000/docs/file/logo.png)\r\n\r\n# 概述\r\n\r\n轻量易用的授权基座smart，提供一个非官方RBAC授权模型的基础开箱即用中台管理服务。\r\n\r\n支持响应码自定义国际化、高粒度权限分配、接口路由配置和实时生效，同时支持部门数据权限横向访问限制。\r\n\r\n本系统的全部接口采用【POST】【application/json】方式传输数据。\r\n\r\n除开放接口以外的其他接口均需要通过【ApiKeyAuth:请求头[Token]】完成鉴权。\r\n\r\n系统技术栈：Golang、VueNext、MySQL、Redis、Gin、ArcoDesign\r\n\r\n# 接口文档\r\n\r\n轻度依赖集成，通过静态HTML加载yaml文件进行打开接口文档。\r\n\r\n该接口文档提供Swagger[支持调试]和ReDoc[阅读增强]两个版本。\r\n\r\n[Swagger[支持调试]：/docs/swagger/index.html](http://localhost:8000/docs/swagger/index.html) 丨 [ReDoc[阅读增强]：/docs/redoc/index.html](http://localhost:8000/docs/redoc/index.html)\r\n\r\n## 生成说明\r\n\r\n- 首次集成\r\n\r\ngo install github.com/swaggo/swag/cmd/swag@latest\r\n\r\n- 格式化注释代码\r\n\r\nswag fmt\r\n\r\n- 初始化swagger.yaml文件\r\n\r\nswag init\r\n\r\n- 如果像本工程一样依赖了api.md文件描述项目\r\n\r\nswag init --md .\r\n\r\n\r\n- 删除多余的生成\r\n\r\nrm .\\docs\\docs.go\r\n\r\n删除docs.go是因为本项目中并未采用下述依赖来集成。\r\n\r\n而是通过HTML模板+JS+YAML引入集成，移除docs.go用于避免依赖编译报错。\r\n\r\n- github.com/swaggo/swag\r\n- github.com/swaggo/gin-swagger\r\n- github.com/swaggo/files\r\n\r\n源帮助页：[https://github.com/swaggo/swag/blob/master/README_zh-CN.md](https://github.com/swaggo/swag/blob/master/README_zh-CN.md)\r\n",
        "title": "smart",
        "contact": {
            "name": "smart",
            "url": "https://smart.siteol.com",
            "email": "dev@siteol.com"
        },
...
        }
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/plat/router/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "新建路由接口配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "路由接口"
                ],
                "summary": "路由接口新建",
                "operationId": "AddRouter 路由接口新建",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platModel.RouterAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/router/del": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "路由接口封存处理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "路由接口"
                ],
                "summary": "路由接口封存",
                "operationId": "DelRouter 路由接口封存",
                "parameters": [
                    {
                        "type": "string",
                        "description": "语言，不传默认为zh-CN",
                        "name": "Lang",
                        "in": "header"
                    },
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/baseModel.IdReq"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/router/edit": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "在路由接口分组下编辑路由接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "路由接口"
                ],
                "summary": "路由接口编辑",
                "operationId": "EditRouter 路由接口编辑",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platModel.RouterEditReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/router/get": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "查询路由接口详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "路由接口"
                ],
                "summary": "路由接口详情",
                "operationId": "GetRouter 路由接口详情",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/baseModel.IdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/platModel.RouterGetRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/router/page": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "查询路由接口分页数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "路由接口"
                ],
                "summary": "路由接口分页",
                "operationId": "PageRouter 路由接口分页",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platModel.RouterPageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/baseModel.PageRes"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/platModel.RouterPageRes"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "baseModel.IdReq": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "数据ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "baseModel.PageRes": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "分页数据"
                },
                "total": {
                    "description": "数据总量",
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "baseModel.ResBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "S0000/F0000"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "Success/Fail"
                },
                "unPop": {
                    "description": "免弹窗提示",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "platModel.RouterAddReq": {
            "type": "object",
            "required": [
                "name",
                "reqLogInDb",
                "reqLogPrint",
                "resLogInDb",
                "resLogPrint",
                "serviceCode",
                "type",
                "url"
            ],
            "properties": {
                "name": {
                    "description": "路由名称，用于界面展示，与权限关联",
                    "type": "string",
                    "maxLength": 32,
                    "example": "Login"
                },
                "reqLogInDb": {
                    "description": "请求日志入库 0 启用 1 默认不启用",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "reqLogPrint": {
                    "description": "请求日志打印 0 打印 1 不打印",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "reqLogSecure": {
                    "description": "请求日志脱敏字段，逗号分隔，打印时允许配置",
                    "type": "string",
                    "maxLength": 256,
                    "example": "phone,account,password"
                },
                "resLogInDb": {
                    "description": "响应日志入库 0 启用 1 默认不启用",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "resLogPrint": {
                    "description": "响应日志打印 0 打印 1 不打印",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "resLogSecure": {
                    "description": "响应日志脱敏字段，逗号分隔，打印时允许配置",
                    "type": "string",
                    "maxLength": 256,
                    "example": "name,account,password"
                },
                "serviceCode": {
                    "description": "业务编码（字典），为接口分组",
                    "type": "string",
                    "maxLength": 3,
                    "example": "base"
                },
                "type": {
                    "description": "免授权路由 0 免授权 1 授权",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "url": {
                    "description": "路由地址，后端访问URL，后端不在URL中携带参数，统一Post处理内容",
                    "type": "string",
                    "maxLength": 64,
                    "example": "/auth/login"
                }
            }
        },
        "platModel.RouterEditReq": {
            "type": "object",
            "required": [
                "id",
                "name",
                "reqLogInDb",
                "reqLogPrint",
                "resLogInDb",
                "resLogPrint",
                "serviceCode"
            ],
            "properties": {
                "id": {
                    "description": "数据ID",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "路由名称，用于界面展示，与权限关联",
                    "type": "string",
                    "maxLength": 32,
                    "example": "Login"
                },
                "reqLogInDb": {
                    "description": "请求日志入库 0 启用 1 默认不启用",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "reqLogPrint": {
                    "description": "请求日志打印 0 打印 1 不打印",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "reqLogSecure": {
                    "description": "请求日志脱敏字段，逗号分隔，打印时允许配置",
                    "type": "string",
                    "maxLength": 256,
                    "example": "phone,account,password"
                },
                "resLogInDb": {
                    "description": "响应日志入库 0 启用 1 默认不启用",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "resLogPrint": {
                    "description": "响应日志打印 0 打印 1 不打印",
                    "type": "string",
                    "enum": [
                        "0",
                        "1"
                    ],
                    "example": "0"
                },
                "resLogSecure": {
                    "description": "响应日志脱敏字段，逗号分隔，打印时允许配置",
                    "type": "string",
                    "maxLength": 256,
                    "example": "name,account,password"
                },
                "serviceCode": {
                    "description": "业务编码（字典），为接口分组",
                    "type": "string",
                    "maxLength": 3,
                    "example": "base"
                }
            }
        },
        "platModel.RouterGetRes": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "数据ID",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "路由名称，用于界面展示，与权限关联",
                    "type": "string",
                    "example": "Login"
                },
                "reqLogInDb": {
                    "description": "请求日志入库 0 启用 1 默认不启用",
                    "type": "string",
                    "example": "0"
                },
                "reqLogPrint": {
                    "description": "请求日志打印 0 打印 1 不打印",
                    "type": "string",
                    "example": "0"
                },
                "reqLogSecure": {
                    "description": "请求日志脱敏字段，逗号分隔，打印时允许配置",
                    "type": "string",
                    "example": "phone"
                },
                "resLogInDb": {
                    "description": "响应日志入库 0 启用 1 默认不启用",
                    "type": "string",
                    "example": "0"
                },
                "resLogPrint": {
                    "description": "响应日志打印 0 打印 1 不打印",
                    "type": "string",
                    "example": "0"
                },
                "resLogSecure": {
                    "description": "响应日志脱敏字段，逗号分隔，打印时允许配置",
                    "type": "string",
                    "example": "name"
                },
                "serviceCode": {
                    "description": "业务编码（字典），为接口分组",
                    "type": "string",
                    "example": "base"
                },
                "type": {
                    "description": "免授权路由 0 免授权 1 授权",
                    "type": "string",
                    "example": "0"
                },
                "url": {
                    "description": "路由地址，后端访问URL，后端不在URL中携带参数，统一Post处理内容",
                    "type": "string",
                    "example": "/auth/login"
                }
            }
        },
        "platModel.RouterPageReq": {
            "type": "object",
            "properties": {
                "current": {
                    "description": "当前页",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "单页数量",
                    "type": "integer",
                    "example": 10
                },
                "serviceCode": {
                    "description": "业务编码（字典），为接口分组",
                    "type": "string",
                    "example": "base"
                },
                "type": {
                    "description": "免授权路由 0 免授权 1 授权",
                    "type": "string",
                    "example": "0"
                },
                "url": {
                    "description": "路由地址，后端访问URL，支持模糊查询",
                    "type": "string",
                    "example": "/auth/login"
                }
            }
        },
        "platModel.RouterPageRes": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "数据ID",
                    "type": "integer",
                    "example": 1
                },
                "mark": {
                    "description": "变更标识 0可变更 1禁止变更",
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "description": "路由名称，用于界面展示，与权限关联",
                    "type": "string",
                    "example": "Login"
                },
                "reqLogInDb": {
                    "description": "请求日志入库 0 启用 1 默认不启用",
                    "type": "string",
                    "example": "0"
                },
                "reqLogPrint": {
                    "description": "请求日志打印 0 打印 1 不打印",
                    "type": "string",
                    "example": "0"
                },
                "reqLogSecure": {
                    "description": "请求日志脱敏字段，逗号分隔，打印时允许配置",
                    "type": "string",
                    "example": "phone"
                },
                "resLogInDb": {
                    "description": "响应日志入库 0 启用 1 默认不启用",
                    "type": "string",
                    "example": "0"
                },
                "resLogPrint": {
                    "description": "响应日志打印 0 打印 1 不打印",
                    "type": "string",
                    "example": "0"
                },
                "resLogSecure": {
                    "description": "响应日志脱敏字段，逗号分隔，打印时允许配置",
                    "type": "string",
                    "example": "name"
                },
                "serviceCode": {
                    "description": "业务编码（字典），为接口分组",
                    "type": "string",
                    "example": "base"
                },
                "type": {
                    "description": "免授权路由 0 免授权 1 授权",
                    "type": "string",
                    "example": "0"
                },
                "url": {
                    "description": "路由地址，后端访问URL，后端不在URL中携带参数，统一Post处理内容",
                    "type": "string",
                    "example": "/auth/login"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    },
...
}